Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/html/webappapis/scripting/processing-model-2/body-onerror-compile-error-data-url-expected.txt

CONSOLE MESSAGE: line 1: SyntaxError: Unexpected token ')'

FAIL <body onerror> - compile error in <script src=data:...> assert_equals: first arg expected "string" but got "object"
FAIL <body onerror> - compile error in <script src=data:...> (column) assert_equals: fourth arg expected "number" but got "undefined"
